add: 한재성 week2 풀이 추가
잘못된 커밋 수정
